requirements.txt
asgiref==3.8.1
autopep8==2.3.1
certifi==2024.12.14
charset-normalizer==3.4.1
diff-match-patch==20241021
Django==5.1.4
django-admin-interface==0.29.2
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-filter==24.3
django-import-export==4.3.3
django-silk==5.3.2
djangorestframework==3.15.2
gprof2dot==2024.6.6
idna==3.10
pillow==11.0.0
pycodestyle==2.12.1
python-slugify==8.0.4
requests==2.32.3
sqlparse==0.5.3
tablib==3.7.0
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
